add crr reporting and move out of home
