+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Reports/CSETWebCore.Reports.csproj
@@ -9,7 +9,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <EmbeddedResource Include="Resources\secrets.json" />
+    <None Include="Resources\secrets.json" />
   </ItemGroup>
 
   <ItemGroup>
