+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Reports/Views/Crr/Index.cshtml
@@ -1,0 +1,5 @@
+﻿@*
+    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+*@
+@{
+}
