+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Reports/Views/Shared/_Layout.cshtml
@@ -25,7 +25,7 @@
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="CrrReport">CRR Report</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Crr" asp-action="CrrReport">CRR Report</a>
                         </li>
                     </ul>
                 </div>
